Rename `vm` to `cx`.
