+++ src/value/mod.rs
@@ -10,8 +10,8 @@ use std::os::raw::c_void;
 use std::marker::PhantomData;
 use neon_runtime;
 use neon_runtime::raw;
-use vm::{Context, FunctionContext};
-use vm::internal::Isolate;
+use cx::{Context, FunctionContext};
+use cx::internal::Isolate;
 use result::{NeonResult, Throw, ResultExt};
 use object::{Object, This};
 use object::class::Callback;
@@ -26,7 +26,7 @@ pub(crate) mod internal {
     use std::os::raw::c_void;
     use neon_runtime;
     use neon_runtime::raw;
-    use vm::{CallbackInfo, FunctionContext};
+    use cx::{CallbackInfo, FunctionContext};
     use value::error::convert_panics;
     use value::{JsObject, Handle, Managed};
     use object::class::Callback;
